+++ examples/async/cli/cli_args.py
@@ -11,3 +11,4 @@ listener = Arg("--listener", str, "event listener host address:port, e.g. 192.16
 card = Arg("--card", int, "card number, e.g. 10058400")
 index = Arg("--index", int, "card index, e.g. 17")
 antipassback = Arg("--antipassback", str, "disabled, (1:2);(3:4), (1,3):(2,4), 1:(2,3) or 1:(2,3,4)")
+profile = Arg("--profile", int, "time profile ID, e.g. 29")