+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     - `put-card-record`
     - `get-status-record`
     - `get-event-record`
+    - `get-time-profile-record`
 
 ### Updated
 1. Removed legacy `#yapf` directives.